+++ test/unittest/verifierSpec.mjs.js
@@ -206,6 +206,12 @@ describe("DKIM Verifier [unittest]", function () {
 				expect(res.signatures[0]?.warnings).to.be.an('array').
 					that.deep.includes({ name: "DKIM_SIGWARNING_UNSIGNED_HEADER", params: ["Reply-To"] });
 			});
+			it("Unsigned Reply-To header that is invalid", async function () {
+				const res = await verifyEmlFile("dkim/unsigned_header-reply_to-invalid.eml");
+				expect(res.signatures.length).to.be.equal(1);
+				expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
+				expect(res.signatures[0]?.warnings).to.be.empty;
+			});
 		});
 	});
 });